HMR support for index.js (by extracting logic into app.js)
